Update rentals migration script to have consistent type across foreign keys

+++ migrations/20250522_11_25_rentals.sql
@@ -9,7 +9,7 @@ CREATE TYPE rate_period AS ENUM (
 
 CREATE TABLE rental_plans (
     id BIGSERIAL PRIMARY KEY,
-    owner_id INTEGER REFERENCES users(id),
+    owner_id BIGINT REFERENCES users(id),
     name TEXT NOT NULL,
     rate NUMERIC(10, 2) NOT NULL,
     rate_period rate_period NOT NULL,
@@ -19,10 +19,10 @@ CREATE TABLE rental_plans (
 );
 
 CREATE TABLE rentals (
-    id SERIAL PRIMARY KEY,
-    plan_id INTEGER REFERENCES rental_plans(id),
-    owner_id INTEGER REFERENCES users(id),
-    tenant_id INTEGER REFERENCES users(id),
+    id BIGSERIAL PRIMARY KEY,
+    plan_id BIGINT REFERENCES rental_plans(id),
+    owner_id BIGINT REFERENCES users(id),
+    tenant_id BIGINT REFERENCES users(id),
     start_date TIMESTAMPTZ NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -31,10 +31,10 @@ CREATE TABLE rentals (
 
 
 CREATE TABLE rental_payments (
-    id SERIAL PRIMARY KEY,
-    rental_id INTEGER REFERENCES rentals(id),
-    payer_id INTEGER REFERENCES users(id),
-    amount NUMERIC NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    rental_id BIGINT REFERENCES rentals(id),
+    payer_id BIGINT REFERENCES users(id),
+    amount NUMERIC(10, 2) NOT NULL,
     payment_date TIMESTAMPTZ NOT NULL,
     proof_of_payment TEXT,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
